requirements.txt
certifi
natsort
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.6
aiosignal==1.3.1
aiosqlite==0.19.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
blessed==1.20.0
boto3==1.33.13
botocore==1.33.13
cached-property==1.5.2
cachetools==5.5.0
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
comm==0.1.4
ConfigArgParse==1.7
croniter==1.3.15
cycler==0.11.0
dash==2.15.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateutils==0.6.12
debugpy==1.7.0
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
descartes==1.1.0
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.2.2
fastapi==0.88.0
fastjsonschema==2.21.1
fire==0.7.0
Flask==2.2.5
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
inquirer==2.10.1
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
json5==0.9.16
jsonpointer==3.0.0
jsonschema==4.17.3
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_server_fileid==0.9.3
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lightning==1.9.5
lightning-cloud==0.5.70
lightning-utilities==0.10.1
loguru==0.7.3
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MinkowskiEngine==0.5.4
mistune==3.0.2
multidict==6.0.5
nbclassic==1.1.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.7.0
nest-asyncio==1.6.0
notebook==6.5.7
notebook_shim==0.2.4
numpy==1.21.6
nuscenes-devkit==1.1.10
omegaconf==2.3.0
open3d==0.17.0
opencv-python==4.10.0.84
ordered-set==4.1.0
packaging==24.0
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
prometheus-client==0.17.1
prompt_toolkit==3.0.48
protobuf==4.24.4
psutil==6.1.0
ptyprocess==0.7.0
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.19
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.4
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
python-editor==1.0.4
python-json-logger==3.0.1
python-multipart==0.0.8
pytorch-lightning==1.9.5
pytz==2024.2
PyYAML==5.4.1
pyzmq==26.2.0
readchar==4.0.5
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
s3transfer==0.8.2
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.3
sentry-sdk==2.19.2
setproctitle==1.3.3
Shapely==1.8.5
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.12.1+cu116
torchmetrics==0.11.4
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==1.26.20
uvicorn==0.22.0
volumentations==0.1.8
wandb==0.18.7
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.13
y-py==0.6.2
yarl==1.9.4
ypy-websocket==0.8.4
zipp==3.15.0